19 s inlined id and adjoins

+++ SpatiocyteProcess.cpp
@@ -33,7 +33,7 @@
 
 LIBECS_DM_INIT(SpatiocyteProcess, Process);
 
-String SpatiocyteProcess::getIDString(unsigned short aVoxel) const
+String SpatiocyteProcess::getIDString(unsigned aVoxel) const
 {
   Variable* aVariable(theSpecies[aVoxel]->getVariable());
   return "["+aVariable->getSystemPath().asString()+":"+
@@ -57,9 +57,11 @@ String SpatiocyteProcess::getIDString(Comp* aComp) const
   return "["+aComp->system->getFullID().asString()+"]";
 }
 
+/*
 String SpatiocyteProcess::getIDString(unsigned int id) const
 {
   Variable* aVariable(theSpecies[id]->getVariable());
   return "["+aVariable->getSystemPath().asString()+":"+
     aVariable->getID()+"]["+int2str(id)+"]";
 }
+*/
